refactor: streamline wallet recovery codes component with new structure
This commit refactors the `wallet-recovery-codes` component by replacing inline code with dedicated components for improved readability and maintainability. Key changes include:

- Introduced `RecoveryCodesActions`, `RecoveryCodesComplete`, `RecoveryCodesDisplay`, and `RecoveryCodesWarning` components to encapsulate specific functionalities.
- Simplified the rendering logic by removing redundant code and enhancing the user experience.
- Updated the `INTERFACE_IDS` in `interface-ids.ts` for consistent formatting.

These changes aim to enhance the onboarding process and provide a more cohesive user experience.
